refactor(contacts): group contacts hook
