+++ features/platform/contacts/src/index.ts
@@ -1,6 +1,6 @@
 export * from "./useContacts";
 export * from "./utils/getContactInitial";
-export * from "./useContactsMeContact";
+export * from "./hooks/useContactsMeContact";
 export * from "./utils/getContactAvatarColorClass";
 export * from "./utils/formatMeDisplayName";
 export * from "./utils/resolveMeContactDisplayName";
